+++ public/cpp/actor/actor_logic/actor_example_manage/message/actor_example_manage_handle.cpp
@@ -45,7 +45,7 @@ namespace ngl
 		if (lpactor != nullptr)
 		{
 			playinfo& lplayinfo = m_info[lprecv->m_type][lpactor->id_guid()];
-			lplayinfo.m_createexample = localtime::gettime();
+			lplayinfo.m_createexample = (int32_t)localtime::gettime();
 			lplayinfo.m_actorexampleid = lpactor->id_guid();
 			for (const auto& lpair : lprecv->m_roleids)
 			{
